+++ Sources/Extensions.swift
@@ -67,3 +67,26 @@ extension Comparable {
         return min(max(self, limits.lowerBound), limits.upperBound)
     }
 }
+
+private let radiusOfEarth: Double = 6372797.6
+
+extension CLLocationCoordinate2D {
+    func coordinate(onBearingInRadians bearing: Double, atDistanceInMeters distance: Double) -> CLLocationCoordinate2D {
+        let distRadians = distance / radiusOfEarth // earth radius in meters
+        
+        let lat1 = latitude * .pi / 180
+        let lon1 = longitude * .pi / 180
+        
+        let lat2 = asin(sin(lat1) * cos(distRadians) + cos(lat1) * sin(distRadians) * cos(bearing))
+        let lon2 = lon1 + atan2(sin(bearing) * sin(distRadians) * cos(lat1), cos(distRadians) - sin(lat1) * sin(lat2))
+        
+        return CLLocationCoordinate2D(latitude: lat2 * 180 / .pi, longitude: lon2 * 180 / .pi)
+    }
+}
+
+extension Dictionary {
+    subscript(key: Key, `default` value: Value) -> Value {
+        mutating get { return self[key] ?? { self[key] = value; return value }() }
+        set { self[key] = newValue }
+    }
+}